requirements.txt
sshtunnel==0.4.0
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==2.1.1
jupyterlab
matplotlib==3.9.2
